Hotfix: remove generated binary assets; add manual asset placeholders
