+++ app/internal/layout.tsx
@@ -1,0 +1,13 @@
+import type { Metadata } from 'next';
+import type { ReactNode } from 'react';
+
+export const metadata: Metadata = {
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
+
+export default function InternalLayout({ children }: { children: ReactNode }) {
+  return children;
+}
