--- app/sitemap.ts
@@ -1,15 +1,0 @@
-import type { MetadataRoute } from 'next';
-
-const siteUrl =
-  process.env.NEXT_PUBLIC_SITE_URL ??
-  (process.env.VERCEL_URL ? `https://${process.env.VERCEL_URL}` : 'http://localhost:3000');
-
-const routes = ['', '/map', '/discover', '/about', '/donate', '/stats', '/submit', '/status'];
-
-export default function sitemap(): MetadataRoute.Sitemap {
-  const lastModified = new Date();
-  return routes.map((route) => ({
-    url: `${siteUrl}${route}`,
-    lastModified,
-  }));
-}
