+++ public/README-ASSETS.md
@@ -1,0 +1,11 @@
+# Asset upload checklist
+
+These are intentionally not committed in this PR; upload via GitHub UI later.
+
+Required files:
+- public/brand/cryptopaymap-logo.png (transparent, square, >=512px)
+- public/favicon.ico
+- public/favicon-16x16.png
+- public/favicon-32x32.png
+- public/apple-touch-icon.png
+- public/og.png
