+++ public/robots.txt
@@ -1,0 +1,7 @@
+User-agent: *
+Allow: /
+Disallow: /api/
+Disallow: /_next/
+Disallow: /internal
+
+Sitemap: https://www.cryptopaymap.com/sitemap.xml
